Rename package names #42
