--- src/main/java-intellij/com/daimler/sechub/util/SecHubUtil.java
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: MIT
-package com.daimler.sechub.util;
-
-public class SecHubUtil {
-
-    public static void logError(String error){
-
-    }
-}
