+++ src/main/java-intellij/com/mercedesbenz/sechub/util/SecHubUtil.java
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT
+package com.mercedesbenz.sechub.util;
+
+public class SecHubUtil {
+
+    public static void logError(String error){
+
+    }
+}
