--- src/main/java/com/daimler/sechub/ui/SecHubRootTeeNode.java
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: MIT
-package com.daimler.sechub.ui;
-
-public class SecHubRootTeeNode extends SecHubTreeNode {
-
-	public SecHubRootTeeNode() {
-		super(null);
-	}
-
-	private static final long serialVersionUID = 1L;
-
-}
