--- src/main/java/com/daimler/sechub/ui/SecHubToolWindowUIContext.java
@@ -1,15 +1,0 @@
-// SPDX-License-Identifier: MIT
-package com.daimler.sechub.ui;
-
-import com.daimler.sechub.util.ErrorLog;
-
-import javax.swing.*;
-
-public class SecHubToolWindowUIContext {
-    public JTable findingTable;
-    public JTree callHierarchyTree;
-    public JTable callHierarchyDetailTable;
-    public ErrorLog errorLog;
-    public JLabel cweIdLabel;
-    public Integer currentSelectedCweId;
-}
