--- src/main/java/com/daimler/sechub/util/SimpleStringUtil.java
@@ -1,32 +1,0 @@
-// SPDX-License-Identifier: MIT
-package com.daimler.sechub.util;
-
-/**
- * A class with different utility methods.
- */
-public class SimpleStringUtil {
-
-    /**
-     * This method removes all space characters (spaces, tabs etc.) from a given string.
-     *
-     * @return String without any space characters.
-     */
-    public static String removeAllSpaces(String stringWithSpaces) {
-        String stringWithNoSpaces = stringWithSpaces.replaceAll("\\s+", "");
-
-        return stringWithNoSpaces;
-    }
-
-    /**
-     * Returns a trimmed string
-     *
-     * @param object or null
-     * @return trimmed toString() or "" when given string was null
-     */
-    public static String toStringTrimmed(Object object) {
-        if (object == null) {
-            return "";
-        }
-        return object.toString().trim();
-    }
-}
