+++ src/main/java/com/mercedesbenz/sechub/ui/SecHubRootTeeNode.java
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+package com.mercedesbenz.sechub.ui;
+
+public class SecHubRootTeeNode extends SecHubTreeNode {
+
+	public SecHubRootTeeNode() {
+		super(null);
+	}
+
+	private static final long serialVersionUID = 1L;
+
+}
