+++ src/main/java/com/mercedesbenz/sechub/util/ErrorLog.java
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: MIT
+package com.mercedesbenz.sechub.util;
+
+public interface ErrorLog {
+
+    default void error(String message) {
+        error(message,null);
+    }
+    default void error(String message, Throwable t) {
+        /* just do nothing */
+    }
+
+    default void warn(String s){
+        /* just do nothing */
+    }
+}
