+++ src/main/java/com/mercedesbenz/sechub/util/JTreeUtil.java
@@ -1,0 +1,14 @@
+// SPDX-License-Identifier: MIT
+package com.mercedesbenz.sechub.util;
+
+import javax.swing.JTree;
+
+public class JTreeUtil {
+	
+	public static void expandAllNodes(JTree tree) {
+		for (int i = 0; i < tree.getRowCount(); i++) {
+			tree.expandRow(i);
+		}
+	}
+
+}
