+++ src/main/resources/META-INF/plugin.xml
@@ -16,12 +16,12 @@
 
     <extensions defaultExtensionNs="com.intellij">
         <!-- Add your extensions here -->
-        <toolWindow id="SecHub" icon="/icons/toolWindowSecHub.svg" factoryClass="com.daimler.sechub.window.SecHubToolWindowFactory" anchor="left"/>
+        <toolWindow id="SecHub" icon="/icons/toolWindowSecHub.svg" factoryClass="com.mercedesbenz.sechub.window.SecHubToolWindowFactory" anchor="left"/>
     </extensions>
 
     <actions>
         <!-- Add your actions here -->
-        <action id="com.daimler.sechub.action.SechubImportAction" class="com.daimler.sechub.action.SechubImportAction"
+        <action id="com.mercedesbenz.sechub.action.SechubImportAction" class="com.mercedesbenz.sechub.action.SechubImportAction"
                 text="Import SecHub Report" description="Import an SecHub Report JSON file into IDE">
             <add-to-group group-id="ToolsMenu" anchor="first" />
         </action>
